authorization finished

+++ backend/routes/posts.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const multer = require('multer');
 const Post = require('../models/post');
+const checkAuth = require('../middleware/check-auth');
 
 
 const router = express.Router();
@@ -21,24 +22,32 @@ const storage = multer.diskStorage({
     }
 })
 
-router.post('',multer({storage:storage}).single("image"),(req,res,next)=>{
+router.post('',checkAuth,
+multer({storage:storage}).single("image"),(req,res,next)=>{
     const url=req.protocol+'://'+req.get('host');
-    const post = new Post({
+     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url+'/images/'+req.file.filename
-    });
-    post.save().then(createdPost=>{
+        imagePath: url+'/images/'+req.file.filename,
+        creator:req.body.creator
+    })
+    post.save()
+    .then(createdPost=>{
         res.status(201).json({
             message:'Post added successfully',
             post: {
                 id: createdPost._id,
                 title: createdPost.title,
                 content: createdPost.content,
-                imagePath: createdPost.imagePath
+                imagePath: createdPost.imagePath,
+                creator: createdPost.creator
             }
         });
-    });
+    }).catch(error => {
+        res.status(500).json({
+          message: "Creating a post failed!"
+        });
+      });
 });
 
 
@@ -71,7 +80,8 @@ router.get('/:id',(req,res,next)=>{
     })
 })
 
-router.put('/:id',multer({storage:storage}).single("image"),(req,res,next)=>{
+router.put('/:id',checkAuth,
+multer({storage:storage}).single("image"),(req,res,next)=>{
     let imagePath
     if(req.file){
         const url=req.protocol+'://'+req.get('host');
@@ -83,20 +93,30 @@ router.put('/:id',multer({storage:storage}).single("image"),(req,res,next)=>{
         _id:req.body.id,
         title:req.body.title,
         content:req.body.content,
-        imagePath:imagePath
+        imagePath:imagePath,
+        creator:req.body.creator
     })
-    Post.updateOne({_id:req.params.id},post).then(result=>{
+    Post.updateOne({_id:req.params.id , creator: req.body.creator},post).then(result=>{
         console.log(result);
-        res.status(200).json({message:'Update successful!',post:post});
-    });
-})
+        if(result.modifiedCount>0){
+            res.status(200).json({message:'Update successful!',post:post});
+    }
+    else{
+        res.status(401).json({message:'Not authorized!'});
+    }
+})})
 
-router.delete('/:id',(req,res,next)=>{
-    console.log(req.params.id);
-    Post.deleteOne({_id:req.params.id}).then(result=>{
-        console.log(result);
+router.delete('/:id',checkAuth,(req,res,next)=>{
+    Post.deleteOne({_id:req.params.id,creator:req.body.creator}).then(result=>{
+        console.log(result,"7843683");
+        if(result.deletedCount>0){
+            res.status(200).json({message:'Deletion successful!'});
+            }
+            else{
+            res.status(401).json({message:'Not authorized!'});
+            }
     });
-    res.status(200).json({message:'Post deleted!'});
+
 })
 
 module.exports = router;
